requirements.txt
fastapi==0.120.0
filterpy==1.4.5
lap==0.5.12
matplotlib==3.10.7
numpy==2.3.4
opencv_contrib_python==4.10.0.84
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
paddleocr==3.2.0
protobuf==6.33.0
pybind11==3.0.1
python-dotenv==1.2.1
scipy==1.16.2
setuptools==80.9.0
skimage==0.0
SQLAlchemy==2.0.44
torch==2.8.0
ultralytics==8.3.201
uvicorn==0.38.0
